🔨 REFACTOR: move ReaderStreamSidebar component to a separate file
