+++ client/reader/following/main.tsx
@@ -1,9 +1,6 @@
 import config from '@automattic/calypso-config';
 import clsx from 'clsx';
 import { translate } from 'i18n-calypso';
-import { useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-import { Dispatch } from 'redux';
 import AsyncLoad from 'calypso/components/async-load';
 import BloganuaryHeader from 'calypso/components/bloganuary-header';
 import NavigationHeader from 'calypso/components/navigation-header';
@@ -11,20 +8,11 @@ import withDimensions from 'calypso/lib/with-dimensions';
 import ReaderOnboarding from 'calypso/reader/onboarding';
 import SuggestionProvider from 'calypso/reader/search-stream/suggestion-provider';
 import ReaderStream, { WIDE_DISPLAY_CUTOFF } from 'calypso/reader/stream';
-import { RecommendedSite } from 'calypso/state/data-layer/wpcom/read/recommendations/sites';
-import {
-	RecommendedSitesRequestAction,
-	requestRecommendedSites,
-} from 'calypso/state/reader/recommended-sites/actions';
-import { getReaderRecommendedSites } from 'calypso/state/reader/recommended-sites/selectors';
-import { READER_RECENT_SIDEBAR_POPULAR_SITES } from '../follow-sources';
 import Recent from '../recent';
-import ReaderPopularSitesSidebar from '../stream/reader-popular-sites-sidebar';
+import ReaderStreamSidebar from './reader-stream-sidebar';
 import { useFollowingView } from './view-preference';
 import ViewToggle from './view-toggle';
 import './style.scss';
-
-export const RECOMMENDED_SITES_SEED = Math.floor( Math.random() * 10001 );
 
 function FollowingStream( { ...props } ) {
 	const { currentView } = useFollowingView();
@@ -55,44 +43,6 @@ function FollowingStream( { ...props } ) {
 			) }
 			<AsyncLoad require="calypso/lib/analytics/track-resurrections" placeholder={ null } />
 		</>
-	);
-}
-
-function ReaderStreamSidebar(): JSX.Element | null {
-	const dispatch = useDispatch< Dispatch< RecommendedSitesRequestAction > >();
-	const recommendedSites = useSelector( ( state ) => {
-		return getReaderRecommendedSites< RecommendedSite >( state, RECOMMENDED_SITES_SEED ) || [];
-	} );
-
-	useEffect( () => {
-		// Avoid fetching recommended sites if they are already present in the store.
-		if ( recommendedSites.length > 0 ) {
-			return;
-		}
-
-		dispatch( requestRecommendedSites( { seed: RECOMMENDED_SITES_SEED, number: 10 } ) );
-	}, [ dispatch, recommendedSites ] );
-
-	if ( recommendedSites.length === 0 ) {
-		return null;
-	}
-
-	return (
-		<ReaderPopularSitesSidebar
-			followSource={ READER_RECENT_SIDEBAR_POPULAR_SITES }
-			items={ recommendedSites.map( ( s ) => {
-				return {
-					blogId: s.blogId,
-					feed_ID: s.feedId,
-					feed_URL: s.feedUrl,
-					site_name: s.title,
-					site_description: s.description,
-					site_icon: s.icon,
-					url: s.url,
-				};
-			} ) }
-			title={ translate( 'Popular sites' ) }
-		/>
 	);
 }
